自己紹介サイトのファイルをpreparationフォルダに移動
