+++ preparation/templates/member-list.html
@@ -1,0 +1,10 @@
+<!-- base.htmle テンプレートを継承 -->
+<!-- 詳しくは https://wonderwall.hatenablog.com/entry/2018/03/13/213000 -->
+{% extends "base.html" %}
+{% block content %}
+<table>
+    <tr>
+        <td>{{tpl_name}}</td>
+    </tr>
+</table>
+{% endblock %}
